test: split `test:queries`
- covers on-chain queries via `vat-localchain` and `vat-orchestration` (when icq module available)
