+++ multichain-testing/ava.queries.config.js
@@ -1,0 +1,6 @@
+import mainConfig from './ava.config.js';
+
+export default {
+  ...mainConfig,
+  files: ['test/queries/**/*.test.ts'],
+};
