+++ multichain-testing/test/queries/README.md
@@ -1,0 +1,20 @@
+## Start Chains
+
+Runs agoric, cosmoshub, and osmosis with hermes relayers.
+
+```sh
+# start starship with default configuration
+make start
+```
+
+## Run Tests
+
+```sh
+yarn test:queries
+```
+
+## Stop Chains
+
+```sh
+make stop
+```
